Starting to converge towards a fixed conic form. Moved ConeType to conic package.
