+++ src/main/scala/org/scalacvx/atoms/AbsAtom.scala
@@ -2,9 +2,7 @@ package org.scalacvx.atoms
 
 import breeze.linalg.DenseMatrix
 import breeze.numerics.abs
-import org.scalacvx.conic.ConicForm
 import org.scalacvx.dcp._
-import org.scalacvx.structures.SparseMatrix
 
 /**
  * Created by lorenzo on 8/18/15.
@@ -16,7 +14,7 @@ case class AbsAtom(expr:Expression) extends Expression {
   override val curvature: Vexity = ConvexVexity
   override val children = Some(Array(ChildExpression(expr, NonDecreasing * expr.sign)))
 
-  //override val conicForm = ??? // ConicForm()
+  override lazy val canonicalize = ???
 
 
 }
