--- src/main/scala/org/scalacvx/conic/ConicObject.scala
@@ -1,8 +1,0 @@
-package org.scalacvx.conic
-
-/**
- * Created by lorenzo on 8/20/15.
- */
-class ConicObject {
-
-}
